refactor: refactor CardDetailFetcher
